added home dir and moved files
