+++ components/home/Epoch.tsx
@@ -6,9 +6,9 @@ import { Divider } from "@material-ui/core";
 import Typography from "@material-ui/core/Typography";
 import { useQuery } from "@apollo/client";
 import { useRouter } from "next/router";
-import MiddleEllipsis from './misc/MiddleEllipsis'
-import ComponentLoader from './misc/ComponentLoader';
-import ErrorMessage from './misc/ErrorMessage';
+import MiddleEllipsis from '../misc/MiddleEllipsis'
+import ComponentLoader from '../misc/ComponentLoader';
+import ErrorMessage from '../misc/ErrorMessage';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { useTheme } from '@material-ui/core/styles';
 import getConfig from 'next/config'
@@ -16,8 +16,8 @@ import {
     PieChart, Pie, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer, Sector, Cell, Line
 } from 'recharts';
 import Card from "@material-ui/core/Card";
-import { GET_BLOCK } from './query/Block'
-import { GET_EPOCH } from "./query/Epoch"
+import { GET_BLOCK } from '../query/Block'
+import { GET_EPOCH } from "../query/Epoch"
 import EpochCountdown from "./EpochCountdown"
 
 const useStyles = makeStyles((theme: Theme) =>
