+++ components/home/EpochCountdown.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
 import Typography from "@material-ui/core/Typography";
 import { useQuery } from "@apollo/client";
-import ComponentLoader from './misc/ComponentLoader';
-import ErrorMessage from './misc/ErrorMessage';
-import { GET_BLOCK } from './query/Block'
-import { GET_EPOCH } from "./query/Epoch"
-import { GET_CHAIN } from "./query/Chain"
+import ComponentLoader from '../misc/ComponentLoader';
+import ErrorMessage from '../misc/ErrorMessage';
+import { GET_BLOCK } from '../query/Block'
+import { GET_EPOCH } from "../query/Epoch"
+import { GET_CHAIN } from "../query/Chain"
 
 
 const EpochCountdown = () => {
@@ -38,31 +38,34 @@ const EpochCountdown = () => {
     let blockNow = currentBlock
     let remainingTime = (lastBlockInEpoch - blockNow) * averageTimeOfBlock;
     let remainingTime2 = (data.epoch.lastBlockNumberForEpoch - blockProposer.data.blocks.blocks[0].number)
-    
 
+   
     let x = setInterval(function () {
+
         let hours = Math.floor((remainingTime % (60 * 60 * 24)) / (60 * 60));
         let minutes = Math.floor((remainingTime % (60 * 60)) / (60));
         let seconds = Math.floor((remainingTime % (60)));
 
         let hrs = document.getElementById("hours") as HTMLInputElement
-        hrs.innerHTML = (hours).toString();
+        hrs ? hrs.innerHTML = (hours).toString() : '0';
         let mins = document.getElementById("minutes") as HTMLInputElement
-        mins.innerHTML = (minutes).toString();
+        mins ? mins.innerHTML = (minutes).toString() : '0';
         let secs = document.getElementById("seconds") as HTMLInputElement
-        secs.innerHTML = (seconds).toString();
+        secs ? secs.innerHTML = (seconds).toString() : '0';
 
         // setBlock(block - 1)
         // blockNow--
-        currentBlock = blockProposer.data.blocks.blocks[0].number - 1
+        //currentBlock = blockProposer.data.blocks.blocks[0].number - 1
+        remainingTime = remainingTime - 1;
 
-    }, 5001);
+
+    }, 1000);
 
     return (
         <>
-            <span id="hours"> </span> <span> h </span>
-            <span id="minutes"> </span> <span> m </span>
-            <span id="seconds"> </span> <span> s </span>
+            <span id="hours">{"0"}</span> <span> h </span>
+            <span id="minutes">{"0"}</span> <span> m </span>
+            <span id="seconds">{"0"}</span> <span> s </span>
         </>
     )
 }
