+++ pages/index.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
 import Layout from "../components/Layout";
-import ChartData from "../components/ChartData";
+import ChartData from "../components/home/ChartData";
 import LatestBlocks from "../components/block/LatestBlocks";
 import LatestTransactions from "../components/transaction/LatestTransactions";
 import Grid from "@material-ui/core/Grid";
 import Ledger from "../components/ledger/Ledger";
-import TokenPrice from "../components/TokenPrice";
-import Epoch from "../components/Epoch";
-import ValidatorsGroups from "../components/ValidatorsGroups";
+import TokenPrice from "../components/home/TokenPrice";
+import Epoch from "../components/home/Epoch";
+import ValidatorsGroups from "../components/home/ValidatorsGroups";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
